Extract HeaderComments helper from SpreadsheetSchema
Move applyHeaderComments body into schema/internal/HeaderComments.
SpreadsheetSchema sheds POI Drawing / Comment / ClientAnchor /
CreationHelper imports and the COMMENT_BOX constants. Writers invoke
HeaderComments.apply directly.

Tighten javadoc on getLeafHeaderRow and forEachHeaderCell for
consistency with sibling schema classes (short single/multi-line,
no <p> paragraphs).
